moved Calculator.js  into its own component folder along with its own newly created Calculator.test.js file; updated App.test.js according to code replacement from App.js to Calculator.js; added UserList.test.js file; added and configured Apollo-client packages
